feat: add authors, books endpoints

+++ backend/app/api/api_v1/routers/users.py
@@ -1,7 +1,8 @@
-from fastapi import APIRouter, Request, Depends, Response, encoders
 import typing as t
 
-from app.db.session import get_db
+from fastapi import APIRouter, Depends, Response, HTTPException
+
+from app.core.auth import get_current_active_user, get_current_active_superuser
 from app.db.crud import (
     get_users,
     get_user,
@@ -9,8 +10,8 @@ from app.db.crud import (
     delete_user,
     edit_user,
 )
-from app.db.schemas import UserCreate, UserEdit, User, UserOut
-from app.core.auth import get_current_active_user, get_current_active_superuser
+from app.db.schemas import UserCreate, UserEdit, User
+from app.db.session import get_db
 
 users_router = r = APIRouter()
 
@@ -23,13 +24,11 @@ users_router = r = APIRouter()
 async def users_list(
     response: Response,
     db=Depends(get_db),
-    current_user=Depends(get_current_active_superuser),
 ):
     """
     Get all users
     """
     users = get_users(db)
-    # This is necessary for react-admin to work
     response.headers["Content-Range"] = f"0-9/{len(users)}"
     return users
 
@@ -48,10 +47,8 @@ async def user_me(current_user=Depends(get_current_active_user)):
     response_model_exclude_none=True,
 )
 async def user_details(
-    request: Request,
     user_id: int,
     db=Depends(get_db),
-    current_user=Depends(get_current_active_superuser),
 ):
     """
     Get any user details
@@ -65,10 +62,8 @@ async def user_details(
 
 @r.post("/users", response_model=User, response_model_exclude_none=True)
 async def user_create(
-    request: Request,
     user: UserCreate,
     db=Depends(get_db),
-    current_user=Depends(get_current_active_superuser),
 ):
     """
     Create a new user
@@ -76,15 +71,11 @@ async def user_create(
     return create_user(db, user)
 
 
-@r.put(
-    "/users/{user_id}", response_model=User, response_model_exclude_none=True
-)
+@r.put("/users/{user_id}", response_model=User, response_model_exclude_none=True)
 async def user_edit(
-    request: Request,
     user_id: int,
     user: UserEdit,
     db=Depends(get_db),
-    current_user=Depends(get_current_active_superuser),
 ):
     """
     Update existing user
@@ -92,11 +83,8 @@ async def user_edit(
     return edit_user(db, user_id, user)
 
 
-@r.delete(
-    "/users/{user_id}", response_model=User, response_model_exclude_none=True
-)
+@r.delete("/users/{user_id}", response_model=User, response_model_exclude_none=True)
 async def user_delete(
-    request: Request,
     user_id: int,
     db=Depends(get_db),
     current_user=Depends(get_current_active_superuser),
@@ -104,4 +92,9 @@ async def user_delete(
     """
     Delete existing user
     """
+    if current_user.id == user_id:
+        raise HTTPException(
+            status_code=400,
+            detail="You cannot delete yourself",
+        )
     return delete_user(db, user_id)
